+++ R/make-newdata.R
@@ -113,25 +113,25 @@ combine_df <- function(...) {
 
 #' Construct a data frame suitable for prediction
 #'
-#' Given a data set, returns a data frame type object that can be used
+#' Given a data set, returns a data set that can be used
 #' as \code{newdata} argument in a call to \code{predict} and similar functions.
-#' The function is particularly usefull in combination with one of the
+#' The function is particularly useful in combination with one of the
 #' \code{add_*} functions, e.g., \code{\link{add_term}}, \code{\link{add_hazard}},
 #' etc.
 #'
 #' @rdname newdata
 #' @aliases make_newdata
 #' @inheritParams sample_info
-#' @param ... Covariate specifications (expressions) that will be evaulated
-#' by looking for variables in \code{x}. Must be of form \code{z = f(z)}
+#' @param ... Covariate specifications (expressions) that will be evaluated
+#' by looking for variables in \code{x}. Must be of the form \code{z = f(z)}
 #' where \code{z} is a variable in the data set \code{x} and \code{f} a known
-#' function that can be usefully applied to \code{age}. See examples below.
+#' function that can be usefully applied to \code{z}. See examples below.
 #' @import dplyr
 #' @importFrom checkmate assert_data_frame assert_character
 #' @importFrom purrr map cross_df
-#' @details Depening on the class of \code{x}, mean or modus values will be
+#' @details Depending on the class of \code{x}, mean or modus values will be
 #' used for variables not specified in ellipsis. If x is an object that inherits
-#' from class \code{ped}, useful data set completion will be performed depending
+#' from class \code{ped}, useful data set completion will be attempted depending
 #' on variables specified in ellipsis.
 #' @examples
 #' tumor %>% make_newdata()
@@ -221,7 +221,7 @@ make_newdata.ped <- function(x, ...) {
     mutate(
       intlen = .data$tend - .data$tstart,
       offset = log(.data$tend - .data$tstart),
-      ped_status = 1)
+      ped_status = 0)
 
 }
 
@@ -247,15 +247,7 @@ make_newdata.fped <- function(x, ...) {
     select(one_of(setdiff(names(x),cumu_vars))) %>%
     unfped() %>% make_newdata(...)
 
-  rest   <- x %>% select(-one_of(c(setdiff(colnames(ndf), "intlen"),
-    unlist(attr(x, "func_mat_names")), int_names)))
-  if (ncol(rest) > 0) {
-    si <- sample_info.data.frame(rest)
-  } else {
-    si <- NULL
-  }
-
-  out_df <- do.call(combine_df, compact(list(si, cumu_smry, ndf)))
+  out_df <- do.call(combine_df, compact(list(cumu_smry, ndf)))
   int_df <- int_info(attr(x, "breaks"))
   suppressMessages(
     out_df <- right_join(int_df, out_df) %>%
@@ -265,7 +257,7 @@ make_newdata.fped <- function(x, ...) {
   # out_df <- adjust_time_vars(out_df, x, dot_names)
 
   ## adjust lag-lead indicator
-  # out_df <- adjust_ll(out_df, x)s
+  out_df <- adjust_ll(out_df, x)
 
   out_df
 
@@ -329,25 +321,25 @@ adjust_ll <- function(out_df, data) {
 }
 
 # All variables that represent follow-up time should have the same values
-adjust_time_vars <- function(out_df, data, dot_names) {
+# adjust_time_vars <- function(out_df, data, dot_names) {
 
-  time_vars <- c("tend",
-    grep(attr(data, "time_var"), unlist(attr(data, "func_mat_names")), value=TRUE))
-  time_vars_dots <- c(grep("tend", dot_names, value=TRUE),
-    grep(attr(data, "time_var"), dot_names, value=TRUE))
-  if(length(time_vars_dots) == 0) {
-    time_vars_dots <- "tend"
-  } else {
-    if(length(time_vars_dots) > 1) {
-      warning(paste0("Only one of ", paste0(time_vars_dots, collapse=", "),
-        "must be specified. Only the first one will be used!"))
-      time_vars_dots <- time_vars_dots[1]
-    }
-  }
-  for(i in setdiff(time_vars, time_vars_dots)) {
-    out_df[[i]] <- out_df[[time_vars_dots]]
-  }
+#   time_vars <- c("tend",
+#     grep(attr(data, "time_var"), unlist(attr(data, "func_mat_names")), value=TRUE))
+#   time_vars_dots <- c(grep("tend", dot_names, value=TRUE),
+#     grep(attr(data, "time_var"), dot_names, value=TRUE))
+#   if(length(time_vars_dots) == 0) {
+#     time_vars_dots <- "tend"
+#   } else {
+#     if(length(time_vars_dots) > 1) {
+#       warning(paste0("Only one of ", paste0(time_vars_dots, collapse=", "),
+#         "must be specified. Only the first one will be used!"))
+#       time_vars_dots <- time_vars_dots[1]
+#     }
+#   }
+#   for(i in setdiff(time_vars, time_vars_dots)) {
+#     out_df[[i]] <- out_df[[time_vars_dots]]
+#   }
 
-  out_df
+#   out_df
 
-}
+# }