+++ tests/testthat/test-cumulative-effects.R
@@ -3,9 +3,50 @@ context("Cumulative effects (of time-dependent covariates)")
 
 test_that("Lag-lead is calculated correctly", {
   LL <- get_laglead(0:2, c(-2, -0.5, 0, 0.5, 2), ll_fun=function(t,tz) {t >= tz})
+  expect_data_frame(LL, nrows = 15L, ncols = 3L)
+  expect_class(LL, "LL_df")
   expect_identical(LL$t, rep(0:2, each = 5))
   expect_identical(LL$tz, rep(c(-2, -0.5, 0, 0.5, 2), times = 3))
-  expect_equal(LL$LL, c(rep(1, 3), rep(0,2), rep(1,4), 0, rep(1,5)))
+  expect_equal(LL$LL, c(rep(0, 5), rep(1,3), rep(0,2), rep(1,4), 0))
+})
+
+test_that("LL helpers and as_ped produce equivalent LL windows", {
+  n <- 1
+  # create data set with variables which will affect the hazard rate.
+  df <- cbind.data.frame(x1 = runif(n, -3, 3)) %>% dplyr::as_tibble()
+  rng_z = function(nz) {rep(1, nz)}
+  # two different exposure times  for two different exposures
+  tz1 <- 1:10
+  tz2 <- -5:5
+   # generate exposures and add to data set
+  df <- df %>% add_tdc(tz1, rng_z) %>% add_tdc(tz2, rng_z)
+
+  # define lag-lead window function
+  ll_fun <- function(t, tz) {t >= tz}
+  ll_fun2 <- function(t, tz) {t >= tz + 2 & t <= tz + 2 + 5}
+  # simulate data with cumulative effect
+  sim_df <- sim_pexp(
+   formula = ~ -3.5 -0.5*x1|fcumu(t, tz1, z.tz1, f_xyz=function(t,tz,z){1},
+    ll_fun=function(t, tz) {t >= tz}) +
+    fcumu(t, tz2, z.tz2, f_xyz=function(t,tz,z){1}, ll_fun=function(t, tz) {t >= tz + 2 & t <= tz + 2 + 5}),
+   data = df, cut = 0:10)
+  sim_df$time <- 10
+
+  ped <- as_ped(sim_df, Surv(time, status) ~ .|cumulative(time, z.tz1, tz_var="tz1") +
+    cumulative(time, z.tz2, tz_var="tz2", ll_fun = function(t, tz) {t >= tz + 2 & t <= tz + 2 + 5}), id = "id")
+  LL1 <- ped$LL_tz1[1:10,]
+  LL1.1 <- get_laglead(0:10, 1:10, ll_fun) %>% filter(t!=0) %>% tidyr::spread(tz, LL)
+  expect_equal(as.matrix(LL1.1[,-1]), LL1, check.attributes=FALSE)
+  LL2 <- ped$LL_tz2[1:10,]
+  LL2.2 <- get_laglead(0:10, -5:5, ll_fun2) %>% filter(t!=0) %>% tidyr::spread(tz, LL)
+  expect_equal(as.matrix(LL2.2[,-1]), LL2, check.attributes=FALSE)
+  LL1.2 <- get_laglead(ped) %>% filter(tz_var == "tz1") %>% filter(t!=0) %>%
+    tidyr::spread(tz, LL) %>% select(-1:-2) %>% as.matrix()
+  LL2.2 <- get_laglead(ped) %>% filter(tz_var == "tz2") %>% filter(t!=0) %>%
+    tidyr::spread(tz, LL) %>% select(-1:-2) %>% as.matrix()
+  expect_equal(LL1, LL1.2, check.attributes=FALSE)
+  expect_equal(LL2, LL2.2, check.attributes=FALSE)
+
 })
 
 test_that("Cumulative effects are calculated correctly", {