+++ tests/testthat/test-formula-utils.R
@@ -1,0 +1,12 @@
+context("Formula utility functions")
+
+test_that("Formula utilities work", {
+
+  expect_identical(get_rhs_vars(~ x1 + sqrt(x2)), c("x1", "x2"))
+  expect_identical(get_tdc_vars(~.|cumulativ(z.tz, tz_var="tz")), "z.tz")
+  expect_identical(get_ped_form(Surv(time, status)~.|cumulative(z.tz, tz_var="tz")),
+    Surv(time, status)~. )
+  expect_true(has_lhs(Surv(time, status)~.))
+  expect_false(has_lhs(~.))
+
+})